removed comments and now allows create config files directly with the --root-dir option

+++ open.py
@@ -10,9 +10,6 @@ import sys
 
 try:
     if getattr(sys, 'frozen', False):
-    # If the application is run as a bundle, the PyInstaller bootloader
-    # extends the sys module by a flag frozen=True and sets the app 
-    # path into variable _MEIPASS'.
         application_path = os.path.dirname(sys.executable)
         
     else:
@@ -89,17 +86,26 @@ def config_exist():
         raise click.ClickException(f'Cannot check if config.json exists: {e}')
 
 
-def create_config():
+def create_config(target_dir):
     
     try: 
-        # ask for new root path
-        new_path = click.prompt('Enter the root path', type=click.Path(exists=True, file_okay=False, dir_okay=True))
-    
-        # create a new file
-        with open(config_file_path, 'w') as file:
-            
-            # append new json with the new path
-            json.dump({"ROOT_DIR": new_path}, file, indent=4)
+
+        if target_dir:
+
+            with open(config_file_path, 'w') as file:
+                
+                json.dump({"ROOT_DIR": target_dir}, file, indent=4)
+
+        else:
+
+
+            new_path = click.prompt('Enter the root path', type=click.Path(exists=True, file_okay=False, dir_okay=True))
+
+
+        
+            with open(config_file_path, 'w') as file:
+                
+                json.dump({"ROOT_DIR": new_path}, file, indent=4)
     
     except Exception as e:
         raise click.ClickException(f'creating config file error: {e}')
@@ -111,23 +117,20 @@ def cli(project_path,root_dir):
     """Open project in VScode no more find, cd, code. save your directory then open projects directly"""
     
     
+    
 
 
 
-    # Check if config file exists
     if config_exist():
 
-        # (IMPORTANT) checks if root dir exists
         if not root_dir_exist():
             try:
                 create_config()
             except:
                 raise click.ClickException('Root dir in config does not exist. run the cli again and confirm to automatically reset the config.json file ')
 
-        # get root dir (even if empty)
         ROOT_DIR = get_root_dir() 
 
-        # check if root dir is a valid path
         if not os.path.exists(ROOT_DIR): 
             try:
                 new_path = click.prompt('Enter the root path', type=click.Path(exists=True, file_okay=False, dir_okay=True))
@@ -135,19 +138,15 @@ def cli(project_path,root_dir):
             except:
                 raise click.ClickException('Root dir in the config file is not a valid path. run the cli again to automatically create the root dir')
 
-        # change root dir if root dir option is present
         if root_dir:
             
-            # check if root dir option path exists
             if not os.path.exists(root_dir):
                 raise click.ClickException(f'root dir presented is not a valid path')
 
-            # if exist, update config.json
             update_root_path(root_dir)
             
         
         
-        #check if rootdir+project is valid
         if not os.path.exists(os.path.join(ROOT_DIR, project_path)):
             raise click.ClickException(f'project path cannot be found. Double check your root dir "{ROOT_DIR}". update via the --root-dir if it is wrong')
 
@@ -164,15 +163,16 @@ def cli(project_path,root_dir):
             click.secho(f"Error output (stderr): {e.stderr}", fg='red')
             click.secho(f"Standard output (stdout): {e.stdout}", fg='red')
 
-            
+    elif not config_exist() and root_dir:
+    
+        create_config(root_dir)
+
+        click.secho('config successfully created. run the cli again', fg='green')
         
-    # if config file dont exist, create a new config file
     else:
 
-        # prompt to create a new config file
         if click.confirm('config.json is not found. do you wish to create a new config.json?'):
 
-            # Create a new config file
         
             create_config()
 
@@ -180,7 +180,6 @@ def cli(project_path,root_dir):
             
             
                 
-        # error: there is no config file
         else:
             raise click.ClickException('config.json is not found. run the cli again and confirm to automatically create the config.json file')
